requirements.txt
torch==1.9.0
torchvision==0.10.0
matplotlib==3.4.2
tensorboard==2.5.0
Pillow==8.2.0
pandas==1.2.4
nltk==3.6.2
einops==0.3.0
kaggle==1.5.12
transformers==4.8.2
pytest
